+++ pycobi/__init__.py
@@ -29,7 +29,7 @@
 
 __author__ = "Richard Gast"
 __status__ = "Development"
-__version__ = "1.0.0"
+__version__ = "1.0.1"
 
 from .pycobi import ODESystem, Continuation
 from .utility import (